优化位置操作时的逻辑。
