+++ src/cli/mod.rs
@@ -1,5 +1,5 @@
 pub mod arg;
-pub mod pos;
+pub mod location;
 mod sign;
 mod single_sign;
